Merged PR 229027: Allow CRM URL to be passed via a data attribute
By default msdyn_cilibrary.js assumes document.referrer points to the CRM domain in which it is embedded. However if multiple redirects have caused the referrer to be overwritten, document.referrer would point to the last redirect and hence postmessages intended for the top (CRM) window would fail.

Due to cross-domain restrictions, the widget iframe cannot directly access the top window's location information nor can the top window set any attribute to pass on the domain info to the widget iframe. This leaves us with two options cross-domain communication between the widget iframe and the parent CRM domain window:
  1. Let the postmessage (at least one initialization message) be broadcast to all ('*') domains
  2. Have the widget provider pass the CRM domain when loading the msdyn_lilibrary.js

This patch implements approach (2) since it approach (1) is less secure. The widget provider will have to add the following attributes to the script tag that loads msdyn_cilibrary.js to pass the CRM domain:
  - 'data-cifid: CIFMainLibrary'
  - 'data-crmurl: <CRM domain name>'

For example,

<script type="text/javascript" src="https://crmorg.crm.dynamics.com/webresources/Widget/msdyn_ciLibrary.js" onload="ciLoadDone();" data-crmurl="https://crmorg.crm.dynamics.com" data-cifid="CIFMainLibrary"></script>

For backward compatibility, we will fallback to using document.referrer if the above attributes are not defined.

Related work items: #1257514