Merged PR 230893: Implement event handler for sendkbarticle event
Add a handler for 'KMClickToSend' window event that will be raised by the KB control for sending a KB article into the CIF widget

This change simply retrieves the event parameters and passes them onto the handler, if any, registered by the widget for this event.

This PR also includes fix for 1262783. Remove references to telemetry infra from the widget-side CIF library

Related work items: #1246307, #1262783