Merged PR 305879: Perform legacy CIF notification related checks only if we are not rendering ZFP notification
The legacy CIF notification infra has a check to automatically discard stacked notifications if there are more than 5. This check was kicking in even for ZFP and trying to delete a non-existent DOM element. Move this check below the check for ZFP

Add a catch-all in postMsgWrapper so that any unhandled exception results in the promise getting rejected instead of getting lost

Related work items: #1451949