Merged PR 333068: Do not dequeue if another notification is being processed.
Do not dequeue if another notification is being processed.

Bug 1501580: ZFP client is not able to handle the multiple notifications coming in a batch.

Race condition when notifications come very close to each other. Tried to display a new notification when one was already being displayed. ideally it should have remained in the notification queue and not dequeued to display. In this case, dequeue was being done on the basis of if there is an active notification. Due to a race condition the second notification was dequeued and then the first one was displayed. Meanwhile the second one also tried to display, which failed.
Corrected the dequeuing logic.
Added reject in case notification display fails.

Cherry picked from !332435

Related work items: #1501580