Merged PR 336947: Fix for notes panel already open
Fix for notes panel already open
Bug 1498753: Clicking on notes icon in conversation control, when the notes is already open is throwing error
In case notes panel is already open, instead of rejecting the promise, resolve it. No need to show an error on the client.
Do nothing if the panel is already open.

Related work items: #1498753