Merged PR 337766: Added scrubbed strings for all admin entities and Notification Button Config custom control
Added scrubbed strings for all admin entities and Notification Button Config custom control

This PR contains just the addition of the scrubbed strings for all the metadata and the controls, so that we can start the process of translation. Rest of the code will be merged in when we merge the admin feature branch.

Related work items: #1506977