Merged PR 343954: Bug 1516011: [Arabic][Loc testing][ZFP] 'null' is being appended to the saved notes while saving the notes from the CC
**Null appendend to notes.**

Before it was not checking that if previous **annotationId** was same as current's and it
selects the last annotationid  which contains null/random message in it  and therefore
cause this bug (**null or random previous message are appending in current message**) to happen.

Now we are checking/matching the annotationid and then append the current message on it.

Related work items: #1516011