Merged PR 360836: Added AcceptButtonText and RejectButtontext attributes in NotificationTemplate
Added AcceptButtonText and RejectButtontext attributes in NotificationTemplate entity

Earlier we were storing these strings as well as the Notification button config in a single attribute (msdyn_NotificationButtons) as a JSON. This will impact localization as jsons strings will not be localized. So adding these strings as a seprate attribute (msdyn_AcceptButtonText and msdyn_RejectButtonText) and will store just the config in msdyn_notificationbuttons attribute.

This PR contains just the data schema change. The custom control change and runtime changes for consuming these new attributes will be taken up in a seperate PR.

Related work items: #1468231