Merged PR 360128: Add Scenario Mapping Control
Scenario Mapping Control is used to diaplay and map individual templates (currently Notification and Session) to the top level entity. This entity can be Channel or Workstream. The msdyn_scenario entity is used for stroing the relationship between templates and the top level channel / workstream to which that scenario belongs.

Related work items: #1468233