Merged PR 366182: Set timer value to null if template has the timer unset.
We are now checking the value of **timeout** in NotificationTemplatesInfra.ts file and if it is null/less than or equal to zero/undefined  then we are not populating the **timeoutAction** property of **ret** object and hence we are getting no value of **Wait time** in the notifications. Wait time is only be shown when the value of timeout is greater than 0.

Related work items: #1575536