Merged PR 391897: Added the try catch block for exception in createTab()
Added the try catch block for exception in createTab().

This exception is thrown when no session exists and the createTab() is called for dashboard, then while attempting to set the parameters to the session template the invalid session id error is thrown.

And this is already being logged inside "_**setSessionTemplateParams()**_" before this exception is thrown.

When a session exists the tab creation is done without any exceptions.

Related work items: #1607641