Merged PR 402264: Changed the aria labels of children elements of tablist to tab from region
Changed the aria labels of children elements of tablist as the children elements of tablist should have a role="tab".

As of now I tested it using the Fiddler. Tested the changes with the Narrator and keyboard as well.

Envt:

https://tp2sg832plainorg01.crm10.dynamics.com/
ocqauser6@ocprrd.onmicrosoft.com
<Standard Titan Password>
C2: https://conversationcontroldev.blob.core.windows.net/v-gaucha/widget.html

**Before Fix:**

![image.png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/809367ff-32ca-4750-8b26-ebf8af625c5a/pullRequests/402264/attachments/image%20(2).png)

**After Fix:**
![image.png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/809367ff-32ca-4750-8b26-ebf8af625c5a/pullRequests/402264/attachments/image.png)

The aria label issue for the Session Panel is fixed. The other one that is still showing up should be addressed by the platform team and can be tracked by workitem #1680277

Related work items: #1646664