Merged PR 583599: Store workflowid as flowId to identify flow within solution
In Macro designer, flowId stores the **workflowidunique** of a flow.
![image.png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/65512151-aff7-4b31-9d3f-909da7920dc9/pullRequests/583599/attachments/image.png)

When this flow is exported with a solution and imported with the solution to another environment, the **workflowidunique** of the flow will be changed, so flowId in the designer cannot be used to identify the flow. In order to make solution flow identifiable in another environment, **workflowid** should be used as flowId.

However, **workflowidunique** needs to be used in execution, some logic is added to retrieve **workflowidunique** from flowId.

After code change:
![image (4).png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/65512151-aff7-4b31-9d3f-909da7920dc9/pullRequests/583599/attachments/image%20%284%29.png)
![image (3).png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/65512151-aff7-4b31-9d3f-909da7920dc9/pullRequests/583599/attachments/image%20%283%29.png)
![image (6).png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/65512151-aff7-4b31-9d3f-909da7920dc9/pullRequests/583599/attachments/image%20%286%29.png)

test org:
https://org85350983.crm.dynamics.com/main.aspx?appid=bdac0186-e270-eb11-b0b0-0022480497a4
admin@CRM318854.onmicrosoft.com
suXdwaWV5V

Related work items: #2237400