Merged PR 628126: Update sessionClosed in legacy pane control
### Description
in legacy pane control, close session calls both onSessionSwitched and onSessionClosed. Both of them calls onSessionContextChanged to set isSessionChanged, which controls if the p.panes will render.

in onSessionClosed() it use closedSessionId as newSessionId. Then we use it to decide if it is chat widget session in onSessionContextChanged and return wrong isSessionChanged

this.sessionChangeEventData.prevSessionId = "Not Available";
this.sessionChangeEventData.newSessionId = closedSessionId;

the fix is use prevSessionId to pass closedSessionId and do special handle in onSessionContextChanged.

### Test org
https://ndegerv436org01.crm16.dynamics.com/main.aspx?appid=acbfb7a1-59f7-eb11-94ee-0022486f4411
crmadmin

### Before submitting this PR, please make sure:
- [x] Run a local build
- [x] Tested with patch solution
- [x] Choose squash commit when completing the PR

Related work items: #2389703