Merged PR 658778: Beethoven Agentscript BorderLess Fix
### Description
We shall not define border color as white if we want a border. At most time, the bundle React style overrider the border-color so we did not discover it in the past time. For Beethoven environment, it somehow does not get overrided by the general layer. That's why we cannot see the border.
![image (2).png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/65512151-aff7-4b31-9d3f-909da7920dc9/pullRequests/658778/attachments/image%20%282%29.png)

### Unit/Integration tests
Visual test

### Test org
https://orga7b503fa.crm10.dynamics.com/main.aspx?forceUCI=1&appid=c7b5755a-7b31-ec11-8c65-000d3a127fdf
crmadmin@cs5x5.onmicrosoft.com

https://aurorabapenv5c3ee.crm10.dynamics.com/main.aspx?forceUCI=1&appid=f239785c-7031-ec11-8c64-000d3a5936f8
aurorauser07@capintegration01.onmicrosoft.com

### Visualization
![image.png](https://dev.azure.com/dynamicscrm/b276c3e1-2902-46bd-a686-484157b97f48/_apis/git/repositories/65512151-aff7-4b31-9d3f-909da7920dc9/pullRequests/658778/attachments/image.png)

### Before submitting this PR, please make sure:
- [X] Run a local build
- [X] Tested with patch solution
- [X] Choose squash commit when completing the PR

'bugFix'

Related work items: #2483346